+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Gitaly changelog
 
-UNRELEASED
+v0.71.0
 
 - Implement GetLfsPointers RPC
   https://gitlab.com/gitlab-org/gitaly/merge_requests/543